% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{attach_biomart}
\alias{attach_biomart}
\title{Attach biomart variables to deseq data}
\usage{
attach_biomart(dat, join_key_dat = "EnsemblGeneID",
  join_key_bm = "ensembl_gene_id", bmart_dataset = "mmusculus_gene_ensembl",
  extra_attributes = c())
}
\arguments{
\item{dat}{Tibble to attach biomart results to.}

\item{join_key_dat}{Join key on the dataset}

\item{join_key_bm}{Join key on biomart}

\item{bmart_dataset}{biomart dataset to be used. Defaults to mmusculus_gene_ensembl}

\item{extra_attributes}{Any extra attributes to attach besides the gene name}
}
\value{
Tibble with attached biomart variables
}
\description{
Attach biomart variables to deseq data
}
