% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{deseq_to_excel}
\alias{deseq_to_excel}
\title{Export deseq folder to excel with default read_deseq}
\usage{
deseq_to_excel(path, outputpath, file_pattern = ".tsv$")
}
\arguments{
\item{path}{Path to deseq files}

\item{outputpath}{Path to output Excel files}

\item{file_pattern}{Pattern to identify deseq files.}
}
\description{
Export deseq folder to excel with default read_deseq
}
