% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{read_deseq}
\alias{read_deseq}
\title{Read in DESeq data from TSV file}
\usage{
read_deseq(path, filter_invalid = TRUE)
}
\arguments{
\item{path}{Path to DESeq data}

\item{filter_invalid}{Remove entry if log2Foldchange, pvalue and padj is NA}
}
\value{
Tibble of DESeq data
}
\description{
Read in DESeq data from TSV file
}
