% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{set_significant}
\alias{set_significant}
\title{Add boolean 'Significant' variable to deseq dataset}
\usage{
set_significant(deseq_dat, padj_min = 0.05)
}
\arguments{
\item{deseq_dat}{Tibble of DESeq data}

\item{padj_min}{Double of minimum adjusted pvalue}
}
\value{
Tibble of entries satisfying significance filter
}
\description{
Add boolean 'Significant' variable to deseq dataset
}
