% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Create a volcano plot from DESeq data set}
\usage{
volcano_plot(deseq_dat, label_top_n = 20)
}
\arguments{
\item{deseq_dat}{DESeq dataset to plot}

\item{label_top_n}{Annotate the top n genes of the data set}
}
\value{
ggplot2 object of volcano plot
}
\description{
Create a volcano plot from DESeq data set
}
